fix(sdk-node)!: fail-fast on Propagator creation from config
This changes the handling of reading `propagato.*` from declarative
config. If a propagator name is unknown, the "create" step will now
*throw* instead of warning. The result will be a no-op SDK.

This also refactors internally from getPropagatorFromConfiguration in
utils.ts to createPropagatorFromConfig in create-from-config.ts, FWIW.

Another minor change in that refactor is that createPropagatorFromConfig
will always wrap propagators in a CompositePropagator, even if there is
only one, because the declarative config schema uses "composite" in the
field names and the description includes:

> Configure the propagators in the composite text map propagator. [...]

Refs: https://github.com/open-telemetry/opentelemetry-js/pull/6785
