+++ experimental/packages/opentelemetry-sdk-node/src/start.ts
@@ -21,7 +21,6 @@ import {
   getInstanceID,
   getMeterReadersFromConfiguration,
   getMeterViewsFromConfiguration,
-  getPropagatorFromConfiguration,
   getResourceDetectorsFromConfiguration,
   getResourceFromConfiguration,
   getSpanProcessorsFromConfiguration,
@@ -46,6 +45,7 @@ import { ATTR_SERVICE_INSTANCE_ID } from './semconv';
 import { diagLogLevelFromSeverityNumberConfig } from './diag';
 import {
   createLoggerProviderFromConfig,
+  createPropagatorFromConfig,
   createSpanLimitsFromConfig,
 } from './create-from-config';
 
@@ -140,13 +140,12 @@ function create(
 
     const resource = setupResource(config, sdkOptions);
 
-    const propagator =
-      sdkOptions?.textMapPropagator === null
-        ? null
-        : (sdkOptions?.textMapPropagator ??
-          getPropagatorFromConfiguration(config));
-    if (propagator) {
-      components.propagator = propagator;
+    if (sdkOptions?.textMapPropagator !== undefined) {
+      if (sdkOptions.textMapPropagator !== null) {
+        components.propagator = sdkOptions.textMapPropagator;
+      }
+    } else if (config.propagator) {
+      components.propagator = createPropagatorFromConfig(config.propagator);
     }
 
     if (config.logger_provider) {
