+++ experimental/packages/opentelemetry-sdk-node/src/utils.ts
@@ -73,12 +73,8 @@ import type {
   NameStringValuePairConfigModel,
   HttpTlsConfigModel,
   GrpcTlsConfigModel,
-  TextMapPropagatorConfigModel,
 } from '@opentelemetry/configuration';
-import {
-  mergePropagatorCompositeConfig,
-  mergeResourceAttributesConfig,
-} from '@opentelemetry/configuration';
+import { mergeResourceAttributesConfig } from '@opentelemetry/configuration';
 import type {
   AggregationOption,
   AggregationSelector,
@@ -107,7 +103,6 @@ import type {
 import type { MetricProducer } from '@opentelemetry/sdk-metrics';
 import { BatchLogRecordProcessor } from '@opentelemetry/sdk-logs';
 import * as fs from 'fs';
-import { inspect } from 'util';
 import { createBatchSpanProcessorFromEnv } from './create-from-env';
 
 const RESOURCE_DETECTOR_ENVIRONMENT = 'env';
@@ -338,112 +333,6 @@ export function getPropagatorFromEnv(): TextMapPropagator | null | undefined {
     // null to signal that the default should **not** be used in its place.
     return null;
   } else if (uniquePropagatorNames.length === 1) {
-    return validPropagators[0];
-  } else {
-    return new CompositePropagator({
-      propagators: validPropagators,
-    });
-  }
-}
-
-/**
- * Get a propagator as defined by configuration model from configuration
- */
-export function getPropagatorFromConfiguration(
-  config: ConfigurationModel
-): TextMapPropagator | undefined {
-  if (!config.propagator) {
-    return undefined;
-  }
-
-  const configComposite = mergePropagatorCompositeConfig(
-    config.propagator.composite,
-    config.propagator.composite_list
-  );
-  if (!configComposite) {
-    return undefined;
-  }
-
-  // TextMapPropagator config items are objects with a single key (the name).
-  // Transform this into a more convenient `(name, value)` 2-tuple.
-  //
-  // As well, guard against two cases where the TypeScript type
-  // `TextMapPropagatorConfigModel` does not exactly represent the JSON schema:
-  // 1. `"minProperties": 1, "maxProperties": 1,`
-  // 2. The type allows keys with an `undefined` value, but the JSON schema
-  //    does not.
-  const kvFromItem = (
-    item: TextMapPropagatorConfigModel
-  ): [string, object | null] => {
-    const keys = [];
-    let value = undefined;
-    for (const key of Object.keys(item)) {
-      value = item[key];
-      if (value === undefined) {
-        continue;
-      }
-      keys.push(key);
-    }
-    if (keys.length !== 1) {
-      throw new Error(
-        `invalid "propagator" entry in configuration, there must be exactly one key (with a non-undefined value): ${inspect(item)}`
-      );
-    }
-    return [keys[0], value as object | null];
-  };
-
-  // First pass: handle 'none', remove dupes.
-  const names = new Set();
-  const kvs = [];
-  for (const item of configComposite) {
-    const kv = kvFromItem(item);
-    const k = kv[0];
-    if (names.has(k)) {
-      continue;
-    }
-    names.add(k);
-    kvs.push(kv);
-    if (k === 'none') {
-      return undefined;
-    }
-  }
-
-  // Implementation note: this only contains specification required propagators that are actually hosted in this repo.
-  // Any other propagators (like aws, aws-lambda, should go into `@opentelemetry/auto-configuration-propagators` instead).
-  const propagatorsFactory = new Map<string, () => TextMapPropagator>([
-    ['tracecontext', () => new W3CTraceContextPropagator()],
-    ['baggage', () => new W3CBaggagePropagator()],
-    ['b3', () => new B3Propagator()],
-    [
-      'b3multi',
-      () => new B3Propagator({ injectEncoding: B3InjectEncoding.MULTI_HEADER }),
-    ],
-    [
-      'jaeger',
-      () => {
-        diag.warn(
-          'The Jaeger propagator is deprecated and will be removed in a future release. Use the W3C TraceContext propagator ("tracecontext") instead.'
-        );
-        return new JaegerPropagator();
-      },
-    ],
-  ]);
-
-  const validPropagators: TextMapPropagator[] = [];
-  for (const [name] of kvs) {
-    const propagator = propagatorsFactory.get(name)?.();
-    if (!propagator) {
-      diag.warn(
-        `Propagator "${name}" requested through configuration is unavailable.`
-      );
-      continue;
-    }
-    validPropagators.push(propagator);
-  }
-
-  if (validPropagators.length === 0) {
-    return undefined;
-  } else if (validPropagators.length === 1) {
     return validPropagators[0];
   } else {
     return new CompositePropagator({
